% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{calcStemVolume}
\alias{calcStemVolume}
\title{Calculate stem volume using an equation}
\usage{
calcStemVolume(dt_stem_i, D, H, eq_type = 1, l_c = NULL)
}
\arguments{
\item{dt_stem_i}{dataframe from the predefined table.}

\item{D}{numeric/vector. DBH in cm.}

\item{H}{numeric/vector. Tree height in m.}

\item{eq_type}{numeric. equation type (currently, 1-4 is defined).}

\item{l_c}{list of parameters. If NULL, parameters will be derived from \code{dt_stem_i}.}
}
\value{
calculated stem volume
}
\description{
Calculate stem volume using an equation that corresponds with the equation type.
This function can handle numeric vectors with different DBH (Diameter at Breast Height) classes.
}
