% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{calcStemVolumeAdj}
\alias{calcStemVolumeAdj}
\title{Calculate stem volume based on junction adjustment, for single value.}
\usage{
calcStemVolumeAdj(dt_stem_i, D, H, eq_type = 1, adj = "None")
}
\arguments{
\item{dt_stem_i}{Dataframe from the predefined table.}

\item{D}{Numeric. DBH in cm.}

\item{H}{Numeric. Tree height in m.}

\item{eq_type}{Numeric. equation type (currently, 1-4 is defined).}

\item{adj}{One of \code{None}, \verb{3w}, \verb{5w}, \verb{3w_o}, \verb{5w_o}.}
}
\value{
Calculated stem volume
}
\description{
Calculate stem volume using an equation corresponding with the equation type.
\code{calcStemVolumeAdj()} only accepts single numeric value for D & H.
To handle numeric vector, use \code{\link[=calcStemVolumeMulti]{calcStemVolumeMulti()}}.
}
\details{
This is a wrapper function to apply 3/5- points moving window average for calculating stem volume
when DBH is located at the junction of equations.\cr
VERY SLOW for the large amount of data.

See the following reference for calculation details:\cr
Hosoda et al. (2010):
"Differences between the present stem volume tables and the values of the volume equations, and their correction"
Jpn. J. For. Plann. 44:23-39.
}
\references{
\url{https://doi.org/10.20659/jjfp.44.2_23}
}
