% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{convertRegionToJapanese}
\alias{convertRegionToJapanese}
\title{Convert region name to Japanese
This function converts region name in English to Japanese.}
\usage{
convertRegionToJapanese(name_region)
}
\arguments{
\item{name_region}{Character. Region name in English.}
}
\value{
Character. Region name in Japanese.
}
\description{
Convert region name to Japanese
This function converts region name in English to Japanese.
}
