% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{convertSppToJapanese}
\alias{convertSppToJapanese}
\title{Convert species name to Japanese}
\usage{
convertSppToJapanese(name_spp)
}
\arguments{
\item{name_spp}{Character. Species name in English.}
}
\value{
Character. Species name in Japanese.
}
\description{
This function converts species name in English to Japanese.
}
