% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getRegionName}
\alias{getRegionName}
\title{Get region & name as list}
\usage{
getRegionName()
}
\value{
A list of parameters for \code{\link[=volumeName]{volumeName()}}
}
\description{
Get region & name as list
}
