% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getStemCoefficients}
\alias{getStemCoefficients}
\title{Generate a list of dataframe for parameters}
\usage{
getStemCoefficients()
}
\value{
A list of dataframe for \code{\link[=stemVolume]{stemVolume()}}
}
\description{
Generate dataframe of parameters for equations 1) stem volume, 2) BH form factor, and
3) height form based volume calculation.
}
