% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getStemCoefficientsRound}
\alias{getStemCoefficientsRound}
\title{Generate a list of rounded parameters}
\usage{
getStemCoefficientsRound()
}
\value{
A list of dataframe for \code{\link[=stemVolume]{stemVolume()}}
}
\description{
Generate dataframe of rounded parameters for equations 1) stem volume, 2) BH form factor, and
3) height form based volume calculation. The significant digits for coefficient a of 1) stem volume are adjusted to
those in Table 2 in Hosoda et al. (2010).
Parameters for 2) BH form factor and 3) height form are the same as those in \code{\link[=getStemCoefficients]{getStemCoefficients()}}.
}
