% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{linearImputeVolume}
\alias{linearImputeVolume}
\title{Linear interpolation for volume}
\usage{
linearImputeVolume(dt_stem_i, d1, d2, D, H, eq_type)
}
\arguments{
\item{dt_stem_i}{dataframe from the predefined table.}

\item{d1}{numeric/vector. lower bound DBH in cm.}

\item{d2}{numeric/vector. upper bound DBH in cm.}

\item{D}{numeric/vector. DBH in cm.}

\item{H}{numeric/vector. Tree height in m.}

\item{eq_type}{numeric. equation type (currently, 1-4 is defined).}
}
\value{
calculated stem volume
}
\description{
Linear imputation between 2-diameter & volume.
Both for Volume calculation & linear interpolation.
}
