% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{round2}
\alias{round2}
\title{Round, not to even digit}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{Numeric.}

\item{digits}{integer. digits to round.}
}
\value{
Rounded numeric.
}
\description{
Implement usual rounding, not to even digits.
}
