% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{stem-volume-functions}
\alias{stem-volume-functions}
\title{A list of functions for calculating stem volume}
\description{
A list of functions for calculating stem volume of trees based on DBH, H, and species.
Main functions are \link{stemVolume} and \link{volumeName}.
}
