% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{stemVolumeSingle}
\alias{stemVolumeSingle}
\title{Calculate stem volume (single)}
\usage{
stemVolumeSingle(Name, D, H, list_data = NULL, off_adj = FALSE, ...)
}
\arguments{
\item{Name}{Japanese character for identifying the calculation methods (e.g., \code{東京スギ}).
This is expected from \code{\link[=volumeName]{volumeName()}}.}

\item{D}{Numeric. DBH in cm.}

\item{H}{Numeric. Tree height in m.}

\item{list_data}{A list of data for coefficients. This is expected from \code{\link[=getStemCoefficients]{getStemCoefficients()}}.
If NULL, internally call \code{\link[=getStemCoefficients]{getStemCoefficients()}}.}

\item{off_adj}{Logical. If TRUE, all moving average adjustment for junctions become inactive.}

\item{...}{Additional arguments. \code{list_data} and \code{adjust_to_excel} are supported.}
}
\value{
Calculated stem volume
}
\description{
Calculate stem volume using coefficients & following the methods described in Hosoda et al. (2010).
This function only accepts single numeric value for D & H. To handle numeric vector, use \code{\link[=stemVolume]{stemVolume()}}.\cr
VERY SLOW if this is used for processing large amount of data.

Different from \code{\link[=stemVolume]{stemVolume()}}, this function can turn off the 3/5 points moving average adjustment if preferred.
}
\details{
For details, please see the following reference:\cr
細田ら (2010) 現行立木幹材積表と材積式による計算値との相違およびその修正方法. 森林計画学会誌 44: 23-39.
(Kazuo HOSODA, Yasushi MITSUDA and Toshiro IEHARA 2010:
"Differences between the present stem volume tables and the values of the volume equations, and their correction"
Jpn. J. For. Plann. 44:23-39.)

There are several differences compared with the description in the original article and
computational program developed based on the same article as follows:
\itemize{
\item The significant digits for the parameters of several equations are different
(e.g., \code{函館エゾマツ}, \code{東京広葉樹}, \code{長野カラマツ}).
\item The additional revisions on \code{札幌トドマツ},  \code{高知天然スギ}, and \code{青森アカマツ} are not reflected.
\item The range for 5 points moving average adjustment is different for \code{青森広葉樹} and \code{高知広葉樹}.
\item The error in floating-point numbers can cause the differences in rounding calculation for \code{北海道針葉樹}.
}
}
\references{
\url{https://doi.org/10.20659/jjfp.44.2_23}
}
