% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{volumeNameSingle}
\alias{volumeNameSingle}
\title{Get Region name (single)}
\usage{
volumeNameSingle(Region, Spp, RS = NULL, name_invalid = NULL)
}
\arguments{
\item{Region}{Character. Region/Prefecture name that are listed in \code{RS}.
Please see the original article by Hosoda et al. (2010) and description in the above program.}

\item{Spp}{Character. Species name in Japanese. If it does not match any species in the lists, then \code{広葉樹} is returned.
Please see the original article by Hosoda et al. (2010) and description in the above program.}

\item{RS}{A list of parameters, which is generated by \link{getRegionName}.
List names should be strictly the same as one generated by \link{getRegionName} to derive the correct parameters.}

\item{name_invalid}{Character to determine the behavior when invalid region name is provided for \code{Region}.
If NULL, cause error with message. If character string provided, return that string for invalid ones.}
}
\value{
Name for identifying and calculating stem volume in \link{stemVolume}
}
\description{
Get Region name that is used in \link{stemVolume}.
Only accept single character for Region & Spp. To handle vector, use a wrapper function \link{volumeName}.
}
\details{
Species & Region names are defined following Hosoda et al. (2010) Jpn. J. For. Plann. 44:23-39.
(https://doi.org/10.20659/jjfp.44.2_23) and
"幹材積計算プログラム" developed by FFPRI.
(https://www.ffpri.affrc.go.jp/database/stemvolume/index.html)
}
