% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{getCoefs}
\alias{getCoefs}
\title{Get coefficients for calculating stem volume}
\usage{
getCoefs(dt_stem_i, D)
}
\arguments{
\item{dt_stem_i}{dataframe from the predefined table.}

\item{D}{numeric. DBH in cm.}
}
\value{
list of coefficients
}
\description{
Get coefficients for calculating stem volume
}
