% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{linearImpute}
\alias{linearImpute}
\title{Linear interpolation}
\usage{
linearImpute(d1, v1, d2, v2, d)
}
\arguments{
\item{d1}{numeric/vector. lower bound DBH in cm.}

\item{v1}{numeric/vector. calculated Volume that corresponds with \code{d1}.}

\item{d2}{numeric/vector. upper bound DBH in cm.}

\item{v2}{numeric/vector. calculated Volume that corresponds with \code{d2}.}

\item{d}{numeric/vector. DBH in cm.}
}
\value{
linear interpolated stem volume or other variables
}
\description{
Linear imputation between 2-diameter & volume.
This can be used not only for DBH & V, but for other parameters.
}
