% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{stemVolume}
\alias{stemVolume}
\title{Calculate stem volume}
\usage{
stemVolume(Name, D, H)
}
\arguments{
\item{Name}{Japanese character for identifying the calculation methods (e.g., \code{東京スギ}).
This is expected from \code{volumeName()}.}

\item{D}{numeric. DBH in cm.}

\item{H}{numeric. Tree height in m.}
}
\value{
calculated stem volume
}
\description{
Calculate stem volume using coefficients & following the methods described in Hosoda et al. (2010).
This function accepts numeric vectors for D & H.
}
\details{
For details, please see the following reference:\cr
細田ら (2010) 現行立木幹材積表と材積式による計算値との相違およびその修正方法. 森林計画学会誌 44: 23-39.\cr
(Kazuo HOSODA, Yasushi MITSUDA and Toshiro IEHARA 2010:
"Differences between the present stem volume tables and the values of the volume equations, and their correction"
Jpn. J. For. Plann. 44:23-39.)

There are several differences compared with the description in the original article and
computational program developed based on the same article as follows:
\itemize{
\item 3点移動平均、5点移動平均が複数適用される場合を反映(2024/01/11)
\item 一部の係数の有効数字で係数の時点でずれが発生している(函館エゾマツ、東京広葉樹の62cm、長野カラマツなど)
\item 形状高法は係数を直線補間してるが、範囲外も上限/下限で線形補間している。
\item マクロにおける「札幌トドマツ」、「高知天然スギ」、「青森アカマツ」の修正は反映していない。
\item 「青森広葉樹」、「高知広葉樹」の5点移動平均の当てはめ範囲がマクロ関数と異なる。
\item 「北海道針葉樹」の小数第3位への四捨五入で、round()が偶数丸めこみ(五捨五入)を行うため誤差が生じる場合がある。
(2024/01/14)関数を定義して修正したが、内部的な数値のズレで合わない場合が存在する(e.g. H=39mかつD>=81cm)
}
}
\references{
\url{https://doi.org/10.20659/jjfp.44.2_23}
}
