% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_volume.R
\name{volumeNameSingle}
\alias{volumeNameSingle}
\title{Get Region name (single)}
\usage{
volumeNameSingle(Region, Spp, RS = NULL, ...)
}
\arguments{
\item{Region}{Character. Region/Prefecture name that are listed in \code{RS}.
Please see the original article by Hosoda et al. (2010) and description in the above program.}

\item{Spp}{Character. Species name in Japanese. If it does not match any species in the lists, then \code{広葉樹} is returned.
Please see the original article by Hosoda et al. (2010) and description in the above program.}

\item{RS}{A list of parameters, which is generated by \code{\link[=getRegionName]{getRegionName()}}.
List names should be strictly the same as one generated by \code{\link[=getRegionName]{getRegionName()}} to derive the correct parameters.}

\item{...}{Additional arguments. \code{name_invalid} and \code{stop_if_NA} are supported.

These can be:
\itemize{
\item \code{stop_if_NA}: logical. If TRUE, cause error when NAs/NaNs/NULL or other invalid region name is provided for \code{Region}.
\item \code{name_invalid}: character (or NA/NULL), which is used as return when invalid region name is provided for \code{Region}.
This is not used if \code{stop_if_NA} is \code{TRUE}.
}}
}
\value{
Name for identifying and calculating stem volume in \code{\link[=stemVolume]{stemVolume()}}
}
\description{
Get Region name that is used in \code{\link[=stemVolume]{stemVolume()}}.
Only accept single character for Region & Spp. To handle vector, use a wrapper function \code{\link[=volumeName]{volumeName()}}.
}
\details{
Species & Region names are defined following Hosoda et al. (2010) Jpn. J. For. Plann. 44:23-39.
(https://doi.org/10.20659/jjfp.44.2_23) and
"幹材積計算プログラム" developed by FFPRI.
(https://www.ffpri.affrc.go.jp/database/stemvolume/index.html)
}
